Insights POM setup
The insights POM now has a constructor to instantiate the page.  
It has a couple of objects from webelements.  
Methods that use the object to carry out sendkeys, clear etc.  
A method that clicks in the keyword box and then clears it.

This is up and running when the test runs.  It's gear that innit.