Resolved issue with movetoelement
This was caused by a misplaced curly bracket in the switch statement.  I
have attempted executing the test with both Chromedriver and FireFox
driver and it works.  This proves the issue was not caused by
incompatability between Selenium 3 and FirefoxDriver.  Also it is not
set as a property like the Chromedriver.